refactor: models and controllers

+++ TopMarket/Areas/Admin/Controllers/ProductImageController.cs
@@ -1,43 +1,47 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿using System.Linq;
 using System.Web.Mvc;
-using TopMarket.Models.EntityFramework;
 using TopMarket.Models;
+using TopMarket.Models.EntityFramework;
 
 namespace TopMarket.Areas.Admin.Controllers
 {
-    public class ProductImageController : Controller
-    {
-        private ApplicationDbContext db = new ApplicationDbContext();
-        // GET: Admin/ProductImage
-        public ActionResult Index(int id)
-        {
-            ViewBag.ProductId = id;
-            var items = db.ProductImages.Where(x => x.ProductId == id).ToList();
-            return View(items);
-        }
+	[Authorize(Roles = "Admin")]
+	public class ProductImageController : Controller
+	{
+		private readonly ApplicationDbContext db = new ApplicationDbContext();
 
-        [HttpPost]
-        public ActionResult AddImage(int productId, string url)
-        {
-            db.ProductImages.Add(new ProductImage
-            {
-                ProductId = productId,
-                Image = url,
-                IsDefault = false
-            });
-            db.SaveChanges();
-            return Json(new { Success = true });
-        }
-        [HttpPost]
-        public ActionResult Delete(int id)
-        {
-            var item = db.ProductImages.Find(id);
-            db.ProductImages.Remove(item);
-            db.SaveChanges();
-            return Json(new { success = true });
-        }
-    }
+		// GET: Admin/ProductImage
+		public ActionResult Index(int id)
+		{
+			ViewBag.ProductId = id;
+			var items = db.ProductImages.Where(x => x.ProductId == id).ToList();
+			return View(items);
+		}
+
+		[HttpPost]
+		public ActionResult AddImage(int productId, string url)
+		{
+			db.ProductImages.Add(new ProductImage
+			{
+				ProductId = productId,
+				Image = url,
+				IsDefault = false
+			});
+
+			db.SaveChanges();
+			return Json(new { Success = true });
+		}
+
+		[HttpPost]
+		public ActionResult Delete(int id)
+		{
+			var item = db.ProductImages.Find(id);
+			if (item == null) return Json(new { success = false });
+
+			db.ProductImages.Remove(item);
+			db.SaveChanges();
+
+			return Json(new { success = true });
+		}
+	}
 }
